% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecase.R
\docType{data}
\name{pecase}
\alias{pecase}
\title{PECASE Award Winners}
\format{A data frame with 1708 rows and 4 columns:
\describe{
  \item{year}{the year the award was granted}
  \item{dept}{the governmental department granting the award}
  \item{name}{the name of the award recipient}
  \item{instit}{the awardee's affiliated institution}
}}
\source{
\url{https://en.wikipedia.org/wiki/Presidential_Early_Career_Award_for_Scientists_and_Engineers}
}
\usage{
pecase
}
\description{
A dataset containing a list of all PECASE award winners.
}
\keyword{datasets}
